requirements.txt
Django>=5.0,<6.0
mysqlclient>=2.2
djangorestframework>=3.15.1